Fix bug where padding was subtracted from widget
This caused combo boxes to draw its items before the
panel padding, causing the text to be clipped off.

The code deleted here seems to have been an effort to
solve what 14dd68e solved so it makes sense that these
would conflict. I'm pretty sure 14dd68e is the correct
solution though and this was some old hack.