Add command to draw polyline from float*
Adds a command which passes a pointer to use for a polyline. Skips pushing data to buffer as an array of shorts and back.

Note: Assumes pointer is user managed and lives until frame is complete*
Usage: When an array of floats is already in memory as an array of nk_vec2{x,y}, same call signature as nk_stroke_polyline()

Can improve on programs which have a smaller fixed buffer for commands.