Button behavior `NK_BUTTON_TRIGGER_ON_RELEASE`
This modifies the trigger behavior of buttons to only activate when the
mouse was also over the same button on the mouse button down event.

Side note:
I also modified my version of Nuklear I use to apply the same logic for
hover, because I use Nuklear for touch input devices where traditional
hover behavior tends to be misleading. I left this out for now because
I can imagine that this could be seen as wrong/faulty behavior by some.