Don't draw outside the bounds of a button.
When drawing the border of a button it's done so via nk_stroke_rect().
The problem with this is when a line is stroked, half of the width of
the line is on one side, and the other half is on the other side. In
the case of a button, this results in half of the border being drawn
outside the bounds of the button. This commit pulls in the border
rectangle by half the border width before performing the stroke.