Adds HEADER to the doxygen output
HEADER is the intro API section of the paq'd files, which is the most
critical part of our documentation. Ideally, this file needs to stay the
same style because it needs to be incorporated into the header file;
however, we can make a minor modification to allow doxygen to render it.
with this commit, I change the name to HEADER.c to allow doxygen
processing. This shouldn't have any affect on the generated paq file.
Eventually, I would like to move this HEADER file into a different name
and the overall mainpage of the documentation.