Adds CRC calculation to draw command buffer
Adding a rolling CRC caluclation based off the commands going into the
draw buffer allows for really quick and cheap difference detection. That
is, for an application to know if it needs to nk_convert and paint the
screen again. All that needs to be done now, is to store the previous
CRC value and compare that to the latest value to determine if you
should draw or not.