Top level Makefile Upgraded
The top level makefile now handles the building of nuklear.h, docs, and
demos.

Doing this, not only provides a more familar interface for C
programmers, but allows the build system to become simplified. Using
`make` can take advantage of multiple processes and threading. Next up
in the line will be updating the workflow to take advantage of these
properties.