fix[crc]: CRC calculation work without breaking the demos.
CRC calculations appear to be updating pretty well. however for panels
that get drawn before others, there seems to be no change in CRC for
small updates on hover, like highlighting an object as you mouse over
it. Note that the CRC does change with highlights in the last panel that
is rendered. More investigation is required.