lots of changes around CRC. not really sure what i did.
I think CRC might be nearly dead. not that I don't want it to work, or
that it is impossible to make work; but I think that the current
implementation of Nuklear makes things muddy. There is a lot of repeat
function names and different files tend to handle behavoir from
different things. I don't know. I think there needs to be some clean up.

Also, the way buffers are closed or "finished" is weird. some seem like
they are treated differently than others, yet are of the same type? some
get `finished` others get `closed` some get both. idk what is up.

And the reason this matters is because the CRC needs to evaluate at the
buffer level; but then needs to be transfered to the context level. So
on close of a buffer, we need to bubble up the crc. but when is a buffer
closed? well.... in several different functions for some strange reason.
I am kind of done trying to figure it out.