fix(build): failure do build from nuklear/src/* sources
This fixes an issue, where you would not be able to build Nuklear from
sources (without amalgamated header), caused by nuklear_math/util.c
not declaring some function definitions, due to wrong preprocessor logic.
This fix is an "Option 3" mentioned in following comment...
https://github.com/Immediate-Mode-UI/Nuklear/issues/256#issuecomment-2949725477
... where we define additional macro at definition (e.g. FOO_IMPLEMENTATION)
so the preprocessor will be able to correctly detect it in all cases.

Fixes: https://github.com/Immediate-Mode-UI/Nuklear/issues/256