fix(demo): nuklear_sdl3_renderer now provides NK_DTOA
There was a note inside nuklear_sdl3_renderer saying that SDL3 does not
have "dtoa", so I simulated it's behavior with printf. This helps us
test the actual NK_DTOA preprocessor path. This neat trick is not mine.
I discovered it here [link below] and tweaked a bit to make it work.
https://github.com/Immediate-Mode-UI/Nuklear/issues/699#issuecomment-2519530861